# TasteX/hotel-room-ordering/backend/openapi/hotel-room-ordering.openapi.yaml
openapi: 3.0.3
info:
  title: Hotel Room Ordering API
  version: 1.1.0
  description: COD-only phase. Payment gateway APIs are disabled for now.
servers:
  - url: https://api.tastex.example.com
paths:
  /api/v1/guest/hotels/{hotelCode}/menu:
    get:
      summary: Get menu by hotel QR code route
      parameters:
        - in: path
          name: hotelCode
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Menu fetched
...
  /api/v1/guest/orders/{orderNumber}/status:
    get:
      summary: Get guest order status
      parameters:
        - in: path
          name: orderNumber
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order status
...